changed nextDeltaRs to better integral type

+++ src/QMCDrivers/ContextForSteps.h
@@ -50,7 +50,7 @@ public:
   int get_num_groups() const { return particle_group_indexes_.size(); }
   RandomGenerator_t& get_random_gen() { return random_gen_; }
 
-  void nextDeltaRs(unsigned int num_rs) {
+  void nextDeltaRs(size_t num_rs) {
     // hate to repeat this pattern, this should never resize.
     walker_deltas_.resize(num_rs);
     makeGaussRandomWithEngine(walker_deltas_, random_gen_);
